# vulkan: Optimize mul_mat_vec p021 and nc shaders