@0cc4m 2025-03-22
Wow, great work. That's a good improvement across all my tests.

ggml_vulkan: 0 = NVIDIA GeForce RTX 3090 (NVIDIA) | uma: 0 | fp16: 1 | warp size: 32 | shared memory: 49152 | matrix cores: NV_coopmat2
| model                          |       size |     params | backend    | ngl |          test |           t/s master |               t/s PR |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: | -------------------: |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |        tg4096 |         66.32 ± 0.00 |         82.59 ± 0.00 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |        tg8192 |         51.65 ± 0.00 |         75.71 ± 0.00 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |       tg16384 |         33.57 ± 0.00 |         64.76 ± 0.00 |

ggml_vulkan: 0 = AMD Radeon (TM) Pro VII (RADV VEGA20) (radv) | uma: 0 | fp16: 1 | warp size: 64 | shared memory: 65536 | matrix cores: none
| model                          |       size |     params | backend    | ngl |          test |           t/s master |               t/s PR |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: | -------------------: |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |        tg4096 |         43.01 ± 0.00 |         51.70 ± 0.00 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |        tg8192 |         32.39 ± 0.00 |         45.45 ± 0.00 |

ggml_vulkan: 0 = Intel(R) Arc(tm) A770 Graphics (DG2) (Intel open-source Mesa driver) | uma: 0 | fp16: 1 | warp size: 32 | shared memory: 65536 | matrix cores: none
| model                          |       size |     params | backend    | ngl |          test |           t/s master |               t/s PR |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ------------: | -------------------: | -------------------: |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |        tg4096 |         16.41 ± 0.00 |         20.09 ± 0.00 |
| llama 8B Q4_K - Small          |   4.36 GiB |     8.03 B | Vulkan     |  99 |        tg8192 |         13.53 ± 0.00 |         19.47 ± 0.00 |