@stduhpf 2025-03-22
The tests were finally all passing on the proprietary driver with the commit just before this one 😢