@jeffbolznv 2025-03-22
Is the wave size of 32 expected? Maybe that's not interacting well with the full REQUIRE_FULL_SUBGROUPS bit, and then that could be somehow breaking the subgroupAdd?