@stduhpf 2025-03-22
Tests are still failing when subgroupAdd is disabled.