@netrunnereve 2025-03-22
I've also got f16 tests failing on my RX 470, but I've only got two of them and I'm on RADV. Welp there goes my weekend!

```
  MUL_MAT(type_a=f16,type_b=f32,m=1056,n=1,k=129,bs=[1,1],nr=[4,1],per=[0,2,1,3],v=0): [MUL_MAT] NMSE = 0.002029398 > 0.000500000 FAIL
  MUL_MAT(type_a=f16,type_b=f32,m=1057,n=1,k=129,bs=[1,1],nr=[4,1],per=[0,2,1,3],v=0): [MUL_MAT] NMSE = 0.003505248 > 0.000500000 FAIL
```

This also fails with subgroupadd turned off and this card has worked file with subgroup stuff in the past.

> Note that the outputs still seem perfectly fine and coherent, so this isn't really a problem.

Oftentimes I've seen that the outputs still look fine if the tests fail with a small NMSE. But if you run a perplexity test the results are noticeably worse.