@stduhpf 2025-03-22
> I've also got f16 tests failing on my RX 470, but I've only got two of them and I'm on RADV. Welp there goes my weekend!

I've got 4 failing tests because I have two AMD GPUs. It's 2 fails per GPU.