@stduhpf 2025-03-23
> Maybe try turning off the disablerobustness flag? But i haven't been able to think of a reason there would be an out of bounds access.

I doesn't look like this is it either. :/