@netrunnereve 2025-03-23
> I've got 4 failing tests because I have two AMD GPUs. It's 2 fails per GPU.

I see. In that case I have 4 too as my W8100 is also failing on the same test :disappointed:. My laptop with an Intel integrated GPU passes.