@jeffbolznv 2025-03-23
Try forcing gqa_ratio = 1 at line 4646?