@0cc4m 2025-03-23
If it happens on RADV too, then I probably missed it due to unrelated mesa bugs I'm dealing with on my Radeon VII. Maybe I need to revise my testing setup.